// frontend/src/lib/generated/query_schemas_pb.ts
// @generated by protoc-gen-es v2.6.3 with parameter "target=ts"
// @generated from file query_schemas.proto (package query, syntax proto3)
 

import type { Message } from "@bufbuild/protobuf";
import type {
  GenEnum,
  GenFile,
  GenMessage,
  GenService,
} from "@bufbuild/protobuf/codegenv2";
import {
  enumDesc,
  fileDesc,
  messageDesc,
  serviceDesc,
} from "@bufbuild/protobuf/codegenv2";

/**
 * Describes the file query_schemas.proto.
 */
export const file_query_schemas: GenFile =
  /*@__PURE__*/
  fileDesc(
    "ChNxdWVyeV9zY2hlbWFzLnByb3RvEgVxdWVyeSJHCgRSb2xlEh4KBHR5cGUYASABKA4yEC5xdWVyeS5Sb2xlLlR5cGUiHwoEVHlwZRIICgRVU0VSEAASDQoJQVNTSVNUQU5UEAEiRwoLRGlhbG9nRW50cnkSGQoEcm9sZRgBIAEoCzILLnF1ZXJ5LlJvbGUSDwoHY29udGVudBgCIAEoCRIMCgRkYXRlGAMgASgDIkEKDFF1ZXJ5UmVxdWVzdBIiCgZkaWFsb2cYASADKAsyEi5xdWVyeS5EaWFsb2dFbnRyeRINCgVxdWVyeRgCIAEoCSIhCg1RdWVyeVJlc3BvbnNlEhAKCHJlc3BvbnNlGAEgASgJMkIKDFF1ZXJ5U2VydmljZRIyCgVRdWVyeRITLnF1ZXJ5LlF1ZXJ5UmVxdWVzdBoULnF1ZXJ5LlF1ZXJ5UmVzcG9uc2ViBnByb3RvMw"
  );

/**
 * @generated from message query.Role
 */
export type Role = Message<"query.Role"> & {
  /**
   * @generated from field: query.Role.Type type = 1;
   */
  type: Role_Type;
};

/**
 * Describes the message query.Role.
 * Use `create(RoleSchema)` to create a new message.
 */
export const RoleSchema: GenMessage<Role> =
  /*@__PURE__*/
  messageDesc(file_query_schemas, 0);

/**
 * @generated from enum query.Role.Type
 */
export enum Role_Type {
  /**
   * @generated from enum value: USER = 0;
   */
  USER = 0,

  /**
   * @generated from enum value: ASSISTANT = 1;
   */
  ASSISTANT = 1,
}

/**
 * Describes the enum query.Role.Type.
 */
export const Role_TypeSchema: GenEnum<Role_Type> =
  /*@__PURE__*/
  enumDesc(file_query_schemas, 0, 0);

/**
 * @generated from message query.DialogEntry
 */
export type DialogEntry = Message<"query.DialogEntry"> & {
  /**
   * @generated from field: query.Role role = 1;
   */
  role?: Role;

  /**
   * @generated from field: string content = 2;
   */
  content: string;

  /**
   * @generated from field: int64 date = 3;
   */
  date: bigint;
};

/**
 * Describes the message query.DialogEntry.
 * Use `create(DialogEntrySchema)` to create a new message.
 */
export const DialogEntrySchema: GenMessage<DialogEntry> =
  /*@__PURE__*/
  messageDesc(file_query_schemas, 1);

/**
 * @generated from message query.QueryRequest
 */
export type QueryRequest = Message<"query.QueryRequest"> & {
  /**
   * @generated from field: repeated query.DialogEntry dialog = 1;
   */
  dialog: DialogEntry[];

  /**
   * @generated from field: string query = 2;
   */
  query: string;
};

/**
 * Describes the message query.QueryRequest.
 * Use `create(QueryRequestSchema)` to create a new message.
 */
export const QueryRequestSchema: GenMessage<QueryRequest> =
  /*@__PURE__*/
  messageDesc(file_query_schemas, 2);

/**
 * @generated from message query.QueryResponse
 */
export type QueryResponse = Message<"query.QueryResponse"> & {
  /**
   * @generated from field: string response = 1;
   */
  response: string;
};

/**
 * Describes the message query.QueryResponse.
 * Use `create(QueryResponseSchema)` to create a new message.
 */
export const QueryResponseSchema: GenMessage<QueryResponse> =
  /*@__PURE__*/
  messageDesc(file_query_schemas, 3);

/**
 * @generated from service query.QueryService
 */
export const QueryService: GenService<{
  /**
   * @generated from rpc query.QueryService.Query
   */
  query: {
    methodKind: "unary";
    input: typeof QueryRequestSchema;
    output: typeof QueryResponseSchema;
  };
}> = /*@__PURE__*/ serviceDesc(file_query_schemas, 0);